reorganize boxsizes
