+++ src/linear_fits/slopes_distribution.jl
@@ -1,0 +1,2 @@
+TODO:
+struct AllSlopesDistribution end
